Move promobot commands into kpromo
No content changes outside of package names.
